Step 4: co-locate producer-paired parsers under their producers
Three clean 1:1 producer/parser pairs now live in one file each:
  - SerialWorker (wheel) + WheelEncoder parser  -> io/devices/encoder.py
  - EncoderSerialInterface (treadmill) + TreadmillSource parser
        -> io/devices/treadmill.py
  - PsychoPyDevice + Psychopy parser  -> io/devices/psychopy_device.py

Each device file ends with `Producer.Parser = <ParserClass>` so the
registry resolves the parser via attribute lookup. The deleted
mesofield/datakit/sources/behavior/{wheel,treadmill,psychopy}.py modules
are gone; SOURCE_REGISTRY now keys "wheel"/"treadmill"/"psychopy" off
the producer classes' Parser attribute.

Camera-side pairings (MMCamera ↔ MesoMetadataSource/PupilMetadataSource)
are 1:many (one MMCamera class, two camera-keyed parsers selected by
YAML). They need camera subclasses to nest cleanly, deferred.

Session-level parsers (config/notes/timestamps/dataqueue) and
analysis-output parsers (mesomap/DLC/suite2p/meso means) stay where
they are -- not producer-paired by nature.

Import-cycle fixes that fell out:
  - mesofield/datakit/sources/__init__.py builds the producer-paired
    entries lazily on first lookup. Eager binding would re-enter the
    device modules mid-init.
  - mesofield/datakit/discover.py uses get_source_class()/available_tags()
    instead of `from .sources import SOURCE_REGISTRY` to avoid forcing
    the lazy load at module import time.
  - mesofield/datakit/loader.py defers DEFAULT_SOURCES construction
    (now via __getattr__ + _get_default_sources) for the same reason.
  - encoder.py uses mesofield.utils._logger instead of
    mesofield.datakit._utils._logger; the latter triggers datakit
    __init__ which loops back through sources.

Verified: test_pipeline.py still passes 45/45; the registry resolves
wheel/treadmill/psychopy to their parser classes inside io.devices.
